move validator to tools/validator
