+++ tools/validator/src/main/resources/application.properties
